move hardware code into agv file
moving all the hardware code out of the main file has the goal of
abstracting the underlying hardware. I also decided to remove the class
structure from the agv because there's always a single instance of it,
since it deals directly with unique peripherals. I also thought of
making the agv class configurable (like the ones you find in mbed), but
that would be a hell lot of work, since I wouldn't be able to use the
cube-generated code anymore. Therefore, better to leave the agv code as
plain functions, without a class, so it doesn't give the idea of being
instanciable more than once.