add tim11 for time-base, increase PWM frequency
adding tim11 with interrupts enabled and a period of 1ms to be used as
time base. Also, I'll be using DC motors instead of the servo ones, so I'm
increasing the PWM frequency to 21kz, above the datasheet's 25kz typical
parameter, while maintaing the ARR at 999, keeping a fine control of the
PWM duty cycle.