implement AGV's basic functionality
implemented the functions TurnAngle and FollowLine. I think those are
the bare minimum I'll need to make the robot perform it's basic task.
There's actually many other function prototypes on the header file, but
I may not even implement those, if these simple ones turn out to be
enough to make the project work. The only hardware-related thing that
still MUST be added is the ultrasonic sensor for collision detection.
Also, I still need to calibrate all parameters in order for the maths to
work correctly, better still would be to get an encoder on the wheels,
but that I would have to look into, better if I can do it without the
encoder.