remove graphGenerator, implement worldGraph
My initial idea was to have the worldGraph agnostic to the physical
structure of the map, but that turned out to be way more complicated
than necessary. Now the worldGraph class builds the graph and
understands the notion of directions in the plane, keeping track of the
instructions it gave to the main so those instructions can be completed
or failed, changing the graph appropriately.